Changes to support more functionality around DEseqDataSet
This required asking the user to be explicity about what type of
`assay_type` the bioconductor containers are holding when calling
`facilitate()`.

This can be specified using either a single-string `assay_type`
parameter, which applies the assay type to the first assay matrix
in the bioconductor container (only matters for multi-assay containers,
like ExpressionSet or SummarizedExperiment (MAE coming later)).

If you want to specificy different assay types for different
assay matrices of multiassay containers (SE, ES), then the user would
use the `facilitate,assay_info` parameter which expects a named list
of lists, where the names are the names of the assay matrices in
the container.