Replaces the `assay_name = "cpm"` DESeqDataSet hack with `normalized = "cpm"
Having a "ghost" assay within a FacileDEseqDataSet object turned out
to be too complicated. Now we allow for `normalized` to be something
else besides TRUE/FALSE in `fetch_assay_data.FacileDESeqDataSet()`.

This paves the way for enabling users to pass in custom normalization
functions through this parameter as well, which is probably a better
idea overall as opposed to switching the normalization strategy based
on the assay_type of the given assay.